requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
qrcode==7.4.2
pillow==10.2.0
sqlalchemy
pydantic
passlib[bcrypt]

